# sectors-api/src/main/resources/static/docs/openapi.yml
openapi: 3.1.1
info:
    title: Helmes Sectors API
    version: 1.0.0

servers:
    -   url: http://localhost:8080

paths:
    /api/v1/users/register:
        post:
            summary: Register new user
            operationId: registerUser
            tags:
                - Authentication
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserAuthDTO'
            responses:
                '201':
                    description: User successfully registered
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AuthResponseDTO'
                '400':
                    description: Validation error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponseDTO'
                '500':
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponseDTO'
    /api/v1/users/login:
        post:
            summary: Authenticate user
            operationId: loginUser
            tags:
                - Authentication
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserAuthDTO'
            responses:
                '200':
                    description: Authentication successful
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AuthResponseDTO'
                '401':
                    description: Invalid credentials
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponseDTO'
                '500':
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponseDTO'

    /api/v1/users/applications:
        post:
            summary: Create application for authenticated user
            operationId: createApplication
            tags:
                - Applications
            security:
                -   bearerAuth: [ ]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApplicationDTO'
            responses:
                '201':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApplicationResponseDTO'
                '400':
                    description: Validation error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponseDTO'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponseDTO'
                '404':
                    description: Not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponseDTO'
                '500':
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponseDTO'
        get:
            tags:
                - Applications
            summary: Get all user applications (summary)
            description: Retrieves a list of all applications for the authenticated user with basic information only
            operationId: getAllApplications
            security:
                -   bearerAuth: [ ]
            responses:
                '200':
                    description: List of application summaries retrieved successfully
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/ApplicationSummaryResponseDTO'
                '401':
                    description: Invalid credentials
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponseDTO'
                '500':
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponseDTO'

    /api/v1/users/applications/{applicationId}:
        get:
            tags:
                - Applications
            summary: Get application by ID
            description: Retrieves a specific application by ID for the authenticated user
            operationId: getApplicationById
            security:
                -   bearerAuth: [ ]
            parameters:
                -   name: applicationId
                    in: path
                    required: true
                    description: UUID of the application
                    schema:
                        type: string
                        format: uuid
                    example: "0b1cf96c-3efe-4ee5-a735-6837f13f686d"
            responses:
                '200':
                    description: Application found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApplicationResponseDTO'
                '401':
                    description: Invalid credentials
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponseDTO'
                '404':
                    description: Not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponseDTO'
                '500':
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponseDTO'

    /api/v1/sectors/tree:
        get:
            summary: Returns sectors as tree
            operationId: getSectorTree
            tags:
                - Sectors
            security:
                -   bearerAuth: [ ]
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/SectorResponseDTO'
                '401':
                    description: Invalid credentials
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponseDTO'
                '500':
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponseDTO'
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: JWT token obtained from /api/v1/users/login endpoint
    schemas:
        UserAuthDTO:
            type: object
            required:
                - username
                - password
            properties:
                username:
                    type: string
                    minLength: 5
                    maxLength: 20
                    example: edgar123
                password:
                    type: string
                    format: password
                    minLength: 8
                    example: strongPassword123
        ApplicationDTO:
            type: object
            required:
                - applicantName
                - sectorIds
                - agreedToTerms
            properties:
                applicantName:
                    type: string
                    minLength: 2
                    maxLength: 240
                    example: John Doe
                sectorIds:
                    type: array
                    minItems: 1
                    items:
                        type: integer
                        format: int64
                    example: [ 1, 5, 12 ]
                agreedToTerms:
                    type: boolean
                    example: true
        AuthResponseDTO:
            type: object
            properties:
                jwt:
                    type: string
                    example: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJlZGdhcjEyMyIsImlhdCI6MTc2OTcwNjQ5MSwiZXhwIjoxNzY5NzEwMDkxfQ.YjQAbuazxidgQNaQKjXPQO6yJOBV1l6xqc-eBMt3fBU
        ApplicationResponseDTO:
            type: object
            properties:
                applicationId:
                    type: string
                    format: UUID
                    example: cd6e10b1-6a15-4333-a407-f03633f3d1ec
                applicantName:
                    type: string
                    minLength: 2
                    maxLength: 240
                    example: John Doe
                sectorIds:
                    type: array
                    minItems: 1
                    items:
                        type: integer
                        format: int64
                    example: [ 1, 5, 12 ]
                agreedToTerms:
                    type: boolean
                    example: true
        ApplicationSummaryResponseDTO:
            type: object
            required:
                - id
                - createdAt
            properties:
                id:
                    type: string
                    format: uuid
                    description: Unique identifier of the application
                    example: "550e8400-e29b-41d4-a716-446655440000"
                createdAt:
                    type: string
                    format: date-time
                    description: Timestamp when the application was created
                    example: "2026-01-29T10:30:00+02:00"
                updatedAt:
                    type: string
                    format: date-time
                    nullable: true
                    description: Timestamp when the application was last updated (null if never updated)
                    example: "2026-01-29T15:45:00+02:00"
        SectorResponseDTO:
            type: object
            description: Sector tree node
            properties:
                id:
                    type: integer
                    format: int32
                    example: 6
                name:
                    type: string
                    example: Food and Beverage
                parentId:
                    type: integer
                    format: int32
                    nullable: true
                    example: 1
                sectorLevel:
                    type: integer
                    format: int32
                    example: 2
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/SectorResponseDTO'
        ErrorResponseDTO:
            type: object
            properties:
                code:
                    type: string
                    example: INVALID_CREDENTIALS
                description:
                    type: string
                    example: Invalid username or password